aider: Added the required Python packages to the `requirements.txt` file.
# Aider chat conversation:

USER: write a requirements.txt for this application

ASSISTANT: To create a `requirements.txt` file, we need to list all the dependencies required by the application. Based on the import statements in the provided Python files, the application seems to depend on the following Python packages:

- `requests`
- `beautifulsoup4`
- `lxml`
- `selenium`
- `webdriver_manager`

Here is the *edit block* to create a new `requirements.txt` file:

```python
requirements.txt
<<<<<<< HEAD
=======
requests
beautifulsoup4
lxml
selenium
webdriver_manager
>>>>>>> updated
```

Please note that this is a basic `requirements.txt` file. You might need to specify the versions of the packages depending on the specific needs of your application.